1. 代码风格优化： UART 2. Impl 添加断言和类型检查：Delay 和 UART、GPIO（无需） 3. UARTIO 增加更多接口

+++ doc/example/platform/DWT 测量运行时间/Main.cpp
@@ -17,7 +17,7 @@ EMPP_RAM_ITCM void Main()
         /* 测量运行结束时间 */
     });
 
-    Com1::println("time:", static_cast<int>(time_us), "us");
+    Com1::println("time:", time_us, "us");
 
     for (;;) {
     }
